Move unit test responses to responses folder
