+++ src/services/blocks/BlockService.spec.ts
@@ -10,9 +10,14 @@ import {
 	kusamaRegistry,
 	polkadotRegistry,
 } from '../../test-helpers/registries';
-import { blockHash789629, getBlock, mockApi, mockBlock789629 } from '../mock';
-import * as block789629 from '../mock/data/block789629.json';
-import * as blocks789629Response from './blocks789629.json';
+import {
+	blockHash789629,
+	getBlock,
+	mockApi,
+	mockBlock789629,
+} from '../test-helpers/mock';
+import * as block789629 from '../test-helpers/mock/data/block789629.json';
+import * as blocks789629Response from '../test-helpers/responses/blocks/blocks789629.json';
 import { BlocksService } from './BlocksService';
 
 /**
