+++ src/services/transaction/TransactionMaterialService.spec.ts
@@ -1,7 +1,7 @@
 import { sanitizeNumbers } from '../../sanitize/sanitizeNumbers';
-import { blockHash789629, mockApi } from '../mock';
+import { blockHash789629, mockApi } from '../test-helpers/mock';
+import * as response789629 from '../test-helpers/responses/transaction/material789629.json';
 import { TransactionMaterialService } from './TransactionMaterialService';
-import * as response789629 from './TransactionMaterialService789629.json';
 
 const transactionMaterialService = new TransactionMaterialService(mockApi);
